CAMEL-19058: check StatefulServiceInitialization just once
This avoids both the type checking and the initialization status on the hot path
